Prevent collection of compass calibration data when compass hardware registers are already loaded with hard iron offsets. Need to recompile with use software calcs instead to be able to calibrate. Add Hard and Soft iron matrices for spool configuration with oceanic and gopro camera. Fix bug in setCompassCalibrationSpoolSetup.

+++ src/main_network_code.cpp
@@ -278,7 +278,7 @@ bool setupOTAWebServer(const char* _ssid, const char* _password, const char* lab
         }
 
         AsyncWebServerResponse *response = request->beginChunkedResponse("text/csv", [](uint8_t *buffer, size_t maxLen, size_t index) -> size_t {
-          static uint16_t currentSample = 0;
+          static int currentSample = 0;
 
           // Reset on first chunk
           if (index == 0) {
